tests: put common/replicated functions into a single file
The scan function was copied/pasted in each utility, this moves
it to a a seperate c file, which is built as a library, which then
every utility statically links to, saving maintainance, and build time.
