+++ tests/CMakeLists.txt
@@ -26,20 +26,22 @@ endif()
 if (CMAKE_COMPILER_IS_GNUCC)
 	set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -DIIO_CHECK_RET")
 endif()
+add_library(iio_tests_helper STATIC iio_common.c gen_code.c)
+target_link_libraries(iio_tests_helper iio)
 
 add_executable(iio_genxml  iio_genxml.c   ${GETOPT_C_FILE} ${LIBIIO_RC})
 add_executable(iio_info    iio_info.c     ${GETOPT_C_FILE} ${LIBIIO_RC})
-add_executable(iio_attr    iio_attr.c gen_code.c    ${GETOPT_C_FILE} ${LIBIIO_RC})
+add_executable(iio_attr    iio_attr.c     ${GETOPT_C_FILE} ${LIBIIO_RC})
 add_executable(iio_readdev iio_readdev.c  ${GETOPT_C_FILE} ${LIBIIO_RC})
 add_executable(iio_reg     iio_reg.c      ${GETOPT_C_FILE} ${LIBIIO_RC})
 add_executable(iio_writedev iio_writedev.c  ${GETOPT_C_FILE} ${LIBIIO_RC})
 
 target_link_libraries(iio_genxml  iio)
-target_link_libraries(iio_info    iio)
-target_link_libraries(iio_attr    iio)
-target_link_libraries(iio_readdev iio)
+target_link_libraries(iio_info    iio iio_tests_helper)
+target_link_libraries(iio_attr    iio iio_tests_helper)
+target_link_libraries(iio_readdev iio iio_tests_helper)
 target_link_libraries(iio_reg     iio)
-target_link_libraries(iio_writedev iio)
+target_link_libraries(iio_writedev iio iio_tests_helper)
 
 set(IIO_TESTS_TARGETS iio_genxml iio_info iio_attr iio_readdev iio_reg iio_writedev)
 
@@ -48,7 +50,7 @@ if(PTHREAD_LIBRARIES)
 	project(iio_stresstest C)
 	add_executable(iio_adi_xflow_check iio_adi_xflow_check.c)
 	add_executable(iio_stresstest iio_stresstest.c)
-	target_link_libraries(iio_adi_xflow_check iio ${PTHREAD_LIBRARIES})
+	target_link_libraries(iio_adi_xflow_check iio iio_tests_helper ${PTHREAD_LIBRARIES})
 	target_link_libraries(iio_stresstest iio ${PTHREAD_LIBRARIES})
 	set(IIO_TESTS_TARGETS ${IIO_TESTS_TARGETS} iio_adi_xflow_check iio_stresstest)
