+++ tests/iio_genxml.c
@@ -24,18 +24,13 @@
 #include <stdio.h>
 #include <string.h>
 
+#include "iio_common.h"
+
 #ifndef _WIN32
 #define _strdup strdup
 #endif
 
 #define MY_NAME "iio_genxml"
-
-enum backend {
-	LOCAL,
-	XML,
-	NETWORK,
-	AUTO,
-};
 
 static const struct option options[] = {
 	  {"help", no_argument, 0, 'h'},
@@ -73,7 +68,7 @@ int main(int argc, char **argv)
 	const char *arg_uri = NULL;
 	const char *arg_xml = NULL;
 	const char *arg_ip = NULL;
-	enum backend backend = LOCAL;
+	enum backend backend = IIO_LOCAL;
 
 	while ((c = getopt_long(argc, argv, "+hn:x:u:",
 					options, &option_index)) != -1) {
@@ -82,24 +77,24 @@ int main(int argc, char **argv)
 			usage();
 			return EXIT_SUCCESS;
 		case 'n':
-			if (backend != LOCAL) {
+			if (backend != IIO_LOCAL) {
 				fprintf(stderr, "-x and -n are mutually exclusive\n");
 				return EXIT_FAILURE;
 			}
-			backend = NETWORK;
+			backend = IIO_NETWORK;
 			arg_ip = optarg;
 			break;
 		case 'x':
-			if (backend != LOCAL) {
+			if (backend != IIO_LOCAL) {
 				fprintf(stderr, "-x and -n are mutually exclusive\n");
 				return EXIT_FAILURE;
 			}
-			backend = XML;
+			backend = IIO_XML;
 			arg_xml = optarg;
 			break;
 		case 'u':
 			arg_uri = optarg;
-			backend = AUTO;
+			backend = IIO_AUTO;
 			break;
 		case '?':
 			return EXIT_FAILURE;
@@ -112,11 +107,11 @@ int main(int argc, char **argv)
 		return EXIT_FAILURE;
 	}
 
-	if (backend == AUTO) 
+	if (backend == IIO_AUTO) 
 		ctx = iio_create_context_from_uri(arg_uri);
-	else if (backend == XML)
+	else if (backend == IIO_XML)
 		ctx = iio_create_xml_context(arg_xml);
-	else if (backend == NETWORK)
+	else if (backend == IIO_NETWORK)
 		ctx = iio_create_network_context(arg_ip);
 	else
 		ctx = iio_create_default_context();
